LLM MODELS/TencentCloud/requirements.txt
tencentcloud-sdk-python==3.0.934
pandas>=1.3.0
python-dotenv>=0.19.0
tqdm>=4.62.0